fix: fixed duplicate fields are outputted in extrafieldinfo #1186
